add breakpoints and colours
