+++ src/themes/theme.ts
@@ -1,30 +1,42 @@
+import pixelsToRem from "#/utilities/pixelsToRem";
+
 export const colours = {
-    RED: "#FF0000",
-    BLACK: "#000000", // Added BLACK for the example
+    SYSTEMBLUE: "#0E53F3",
+    MIDNIGHT: "#010101",
+    NARDOGREY: "#5F5F5F",
+    CANVAS: "#FFF",
 };
 
-const FULL = 8;
+const FULL = pixelsToRem(8);
 
 export const spacings = {
-    /* 8px */
-    FULL: `${FULL}px`,
-    /* 4px */
-    HALF: `${FULL / 2}px`,
-    /* 16px */
-    DOUBLE: `${FULL * 2}px`,
-    /* 24px */
-    TRIPLE: `${FULL * 3}px`,
-    /* 32px */
-    QUADRUPLE: `${FULL * 4}px`,
-    /* 40px */
-    QUINTUPLE: `${FULL * 5}px`,
-    /* 48px */
-    SEXTUPLE: `${FULL * 6}px`,
+    /* 0.5rem (8px) */
+    FULL: `${FULL}rem`,
+    /* 0.25rem (4px) */
+    HALF: `${FULL / 2}rem`,
+    /* 1rem (16px) */
+    DOUBLE: `${FULL * 2}rem`,
+    /* 1.5rem (24px) */
+    TRIPLE: `${FULL * 3}rem`,
+    /* 2rem (32px) */
+    QUADRUPLE: `${FULL * 4}rem`,
+    /* 2.5rem (40px) */
+    QUINTUPLE: `${FULL * 5}rem`,
+    /* 3rem (48px) */
+    SEXTUPLE: `${FULL * 6}rem`,
+};
+
+export const breakpoints = {
+    GROUP_0_FLIPPHONE: "320px",
+    GROUP_1_SMARTPHONE: "480px",
+    GROUP_2_TABLET: "770px",
+    GROUP_3_COMPUTER: "1000px",
 };
 
 const theme = {
     colours,
     spacings,
+    breakpoints,
 };
 
 export default theme;
